game.play

+++ making_games_01/game_number_guessing.h
@@ -6,23 +6,23 @@
 
 class Game {
  public:
-    static const uint MAX_NUMBER = 100;
+    static const uint8_t MAX_NUMBER = 100;
 
-    static uint random();
-    static uint random(const uint min, const uint max);
+    static uint8_t random();
+    static uint8_t random(const uint8_t min, const uint8_t max);
 
     Game();
-    explicit Game(const uint number);
+    explicit Game(const uint8_t number);
 
-    const int8_t check_answer(const uint input_num);
+    const int8_t check_answer(const int input_num);
     const bool play();
     const bool game_is_over() const;
     const bool player_won() const;
 
  private:
-    static const uint MAX_TRIES = 6;
+    static const uint8_t MAX_TRIES = 6;
 
-    const uint _get_num() const;
+    const int _get_num() const;
 
     void _msg_answer_range() const;
     void _msg_try_again() const;
@@ -34,10 +34,10 @@ class Game {
     void _cout() const;
     void _cout(const std::string& msg, const bool endl = true) const;
 
-    const uint _number;
-    uint _min = 0;
-    uint _max = MAX_NUMBER;
-    uint _guesses_left = MAX_TRIES;
+    const uint8_t _number;
+    uint8_t _min = 0;
+    uint8_t _max = MAX_NUMBER;
+    uint8_t _guesses_left = MAX_TRIES;
     bool _won;
 };
